requirements.txt
Django==4.1.7
djangorestframework==3.14.0
stripe==9.6.0